Function socket_bench_server::server_stream_handler

pub async fn server_stream_handler<L, I, S, E>(server_listener: L) where
    L: Stream<Item = Result<(I, NetworkAddress), E>> + Unpin,
    I: Future<Output = Result<S, E>> + Send + 'static,
    S: AsyncRead + AsyncWrite + Unpin + Send + 'static,
    E: Error + Send, 

Server side handler for send throughput benchmark when the messages are sent over a simple stream (tcp or in-memory).